LAB7/requirements.txt
pika==1.1.0
psycopg2-binary==2.8.6
requests==2.25.1
beautifulsoup4==4.9.3
